services/tutorial-executor/backend/requirements.txt
aio-pika==9.5.4
aioconsole==0.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiormq==6.8.1
aiosignal==1.3.2
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
astunparse==1.6.3
backoff==2.2.1
beautifulsoup4
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
environs==11.2.1
expecttest==0.2.1
fastapi==0.115.6
fastapi_cors==0.0.6
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.2
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.66.0
greenlet==3.0.1
grpcio==1.68.1
grpcio-status==1.62.3
h11==0.14.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
hypothesis==6.115.5
importlib_metadata==8.5.0
inquirerpy==0.3.4
joblib==1.4.2
jsonlines==4.0.0
kombu==5.4.2
lintrunner==0.12.5
litellm==1.35.32
loguru==0.7.3
lxml==4.9.3
Mako==1.3.8
marshmallow==3.23.2
motor==3.6.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
ninja==1.11.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
openai==1.24.0
optree==0.13.0
pamqp==3.3.0
pfzy==0.3.4
pillow==10.2.0
playwright==1.40.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pyee==11.0.1
pymongo==4.9.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-etcd==0.4.5
redis==5.2.1
regex==2024.11.6
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchelastic==0.2.2
torchvision==0.20.1+cu124
transformers==4.47.1
triton==3.1.0
types-dataclasses==0.6.6
typing_extensions
tzdata==2024.2
uritemplate==4.1.1
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.3
websockets==14.1
yarl==1.18.3
zstandard==0.23.0
